requirements.txt
google-cloud-aiplatform[agent_engines, adk]
google-adk
cloudpickle
pydantic
pyyaml
ipykernel
python-dotenv
google-cloud-bigquery
Flask
gunicorn